[feat] 카드 레이아웃 변경 및 적용
